+++ posthog-server/src/main/java/com/posthog/server/PostHogConfig.kt
@@ -166,7 +166,7 @@ public open class PostHogConfig constructor(
                 encryption = encryption,
                 onFeatureFlags = onFeatureFlags,
                 proxy = proxy,
-                remoteConfigProvider = { config, api, _, _ ->
+                remoteConfigProvider = { config, api, _, _, _ ->
                     PostHogFeatureFlags(
                         config,
                         api,